mcp_demo/pyproject.toml
[project]
name = "mcp-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.12.3",
    "pillow>=11.3.0",
]
